split rule set into its own class
