+++ routes/api.php
@@ -16,6 +16,7 @@ Route::prefix('v1')->name('api.v1.')->namespace('API')->group(function () {
     Route::middleware('auth:api')->group(function () {
         Route::apiResource('shows', 'ShowController');
         Route::patch('shows/{show}/hosts', 'ShowController@changeHosts');
+        Route::put('shows/{show}/submitted', 'ShowController@submit');
 
         Route::get('users', 'UserController@search');
     });
